Update base image to alpine 3.20
There are no new updates being built for 3.18, so update to 3.20 to get new security updates.

The pip installed by this is more picky about updating system dependencies and refuses to install without the extra flag.
Ideally we'd move to install the aws cli v2 which is a direct apk package, but this requires more care to ensure compatibility with CI scripts, so leaving as a v1 CLI for now.